2.2
Pingowanie na chacie, zmiana systemu xp, poprawa dzialania

+++ src/com/rpg/core/commands/CommandSetExp.java
@@ -16,12 +16,12 @@ public class CommandSetExp implements CommandExecutor{
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) 
     {
     	Player player = Bukkit.getPlayer(args[0]);
-    	String messageOne = CoreConfig.infoColor + "Ustawiono ilosc doswiadczenia gracza " + player.getDisplayName() + CoreConfig.infoColor + " na &l" + args[1]  + " ["+ Misc.ExpToLvl(Integer.parseInt(args[1])) +"]";
-    	String messageTwo = CoreConfig.infoColor + "Zmieniono ilosc twojego doswiadczenia na " + CoreConfig.infoColor + "&l" + args[1] + " ["+ Misc.ExpToLvl(Integer.parseInt(args[1])) +"]";
+    	String messageOne = CoreConfig.infoColor + "Ustawiono ilosc doswiadczenia gracza " + player.getDisplayName() + CoreConfig.infoColor + " na &l" + args[1]  + " ["+ Misc.ExpToLvl(Long.parseLong(args[1])) +"]";
+    	String messageTwo = CoreConfig.infoColor + "Zmieniono ilosc twojego doswiadczenia na " + CoreConfig.infoColor + "&l" + args[1] + " ["+ Misc.ExpToLvl(Long.parseLong(args[1])) +"]";
     	((Player)sender).sendMessage(StringManager.Colorize(messageOne));
     	player.sendMessage(StringManager.Colorize(messageTwo));
-    	PlayerManager.setExp(player.getUniqueId().toString(), Integer.parseInt(args[1]));
-    	PlayerManager.UpdateExpBar(player, Integer.parseInt(args[1]));
+    	PlayerManager.setExp(player.getUniqueId().toString(), Long.parseLong(args[1]));
+    	PlayerManager.UpdateExpBar(player, Long.parseLong(args[1]));
     	return true;
     }
 }
