+++ src/com/rpg/core/framework/PlayerInfo.java
@@ -3,7 +3,7 @@ package com.rpg.core.framework;
 public class PlayerInfo {
 	public String name;
 	public String UUID;
-	public int experience;
+	public long experience;
 	public int level;
 	public int money;
 }
